# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project Management System API for Wind/Solar Energy Forecasting Company
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List contracts
      parameters:
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
        - name: province
          in: query
          schema:
            type: ["string", "null"]
        - name: group
          in: query
          schema:
            type: ["string", "null"]
        - name: station
          in: query
          schema:
            type: ["string", "null"]
        - name: salesManager
          in: query
          schema:
            type: ["string", "null"]
        - name: contractType
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractBody"
      responses:
        "201":
          description: Contract created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get contract by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Not found
    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Update contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractBody"
      responses:
        "200":
          description: Contract updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Not found
    delete:
      operationId: deleteContract
      tags: [contracts]
      summary: Delete contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /contracts/{id}/change-logs:
    get:
      operationId: getContractChangeLogs
      tags: [contracts]
      summary: Get change history for a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Change logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContractChangeLog"
  /contracts/summary:
    get:
      operationId: getContractSummary
      tags: [contracts]
      summary: Get contract summary statistics
      parameters:
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractSummary"

  # ============ PAYMENTS (回款) ============
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: contractId
          in: query
          schema:
            type: ["integer", "null"]
        - name: contractNo
          in: query
          schema:
            type: ["string", "null"]
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
        - name: province
          in: query
          schema:
            type: ["string", "null"]
        - name: salesManager
          in: query
          schema:
            type: ["string", "null"]
        - name: overdueOnly
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceBody"
      responses:
        "200":
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Contract:
      type: object
      properties:
        id:
          type: integer
        contractNo:
          type: string
        contractName:
          type: string
        contractType:
          type: string
        status:
          type: string
        customer:
          type: string
        province:
          type: string
        group:
          type: string
        station:
          type: string
        salesManager:
          type: string
        signDate:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        amountWithTax:
          type: number
        amountWithoutTax:
          type: number
        productType:
          type: string
        workOrderNo:
          type: ["string", "null"]
        afterSaleNo:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isSpecial:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - contractNo
        - contractName
        - contractType
        - status
        - customer
        - province
        - group
        - station
        - salesManager
        - amountWithTax
        - amountWithoutTax
        - productType
        - isSpecial
        - createdAt
        - updatedAt
    CreateContractBody:
      type: object
      properties:
        contractNo:
          type: string
        contractName:
          type: string
        contractType:
          type: string
        status:
          type: string
        customer:
          type: string
        province:
          type: string
        group:
          type: string
        station:
          type: string
        salesManager:
          type: string
        signDate:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        amountWithTax:
          type: number
        amountWithoutTax:
          type: number
        productType:
          type: string
        workOrderNo:
          type: ["string", "null"]
        afterSaleNo:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isSpecial:
          type: boolean
      required:
        - contractNo
        - contractName
        - contractType
        - status
        - customer
        - province
        - group
        - station
        - salesManager
        - amountWithTax
        - amountWithoutTax
        - productType
    UpdateContractBody:
      type: object
      properties:
        contractNo:
          type: ["string", "null"]
        contractName:
          type: ["string", "null"]
        contractType:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        customer:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        group:
          type: ["string", "null"]
        station:
          type: ["string", "null"]
        salesManager:
          type: ["string", "null"]
        signDate:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        amountWithTax:
          type: ["number", "null"]
        amountWithoutTax:
          type: ["number", "null"]
        productType:
          type: ["string", "null"]
        workOrderNo:
          type: ["string", "null"]
        afterSaleNo:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isSpecial:
          type: ["boolean", "null"]
    ContractChangeLog:
      type: object
      properties:
        id:
          type: integer
        contractId:
          type: integer
        fieldName:
          type: string
        oldValue:
          type: ["string", "null"]
        newValue:
          type: ["string", "null"]
        changedAt:
          type: string
        changedBy:
          type: ["string", "null"]
      required:
        - id
        - contractId
        - fieldName
        - changedAt
    ContractSummary:
      type: object
      properties:
        totalWithTax:
          type: number
        totalWithoutTax:
          type: number
        count:
          type: integer
        byProvince:
          type: array
          items:
            $ref: "#/components/schemas/DimensionStat"
        byManager:
          type: array
          items:
            $ref: "#/components/schemas/DimensionStat"
        byProductType:
          type: array
          items:
            $ref: "#/components/schemas/DimensionStat"
      required:
        - totalWithTax
        - totalWithoutTax
        - count
        - byProvince
        - byManager
        - byProductType

    # ============ PAYMENT ============
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        contractId:
          type: ["integer", "null"]
        contractNo:
          type: ["string", "null"]
        invoiceNo:
          type: ["string", "null"]
        customer:
          type: string
        province:
          type: string
        station:
          type: string
        salesManager:
          type: string
        invoiceDate:
          type: string
        amountWithTax:
          type: number
        amountWithoutTax:
          type: number
        taxRate:
          type: number
        expectedPaymentDate:
          type: ["string", "null"]
        expectedPaymentAmount:
          type: ["number", "null"]
        actualPaymentDate:
          type: ["string", "null"]
        actualPaymentAmount:
          type: ["number", "null"]
        voidDate:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        outstandingAmount:
          type: number
        isOverdue:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - customer
        - province
        - station
        - salesManager
        - invoiceDate
        - amountWithTax
        - amountWithoutTax
        - taxRate
        - status
        - outstandingAmount
        - isOverdue
        - createdAt
        - updatedAt
    CreateInvoiceBody:
      type: object
      properties:
        contractId:
          type: ["integer", "null"]
        contractNo:
          type: ["string", "null"]
        invoiceNo:
          type: ["string", "null"]
        customer:
          type: string
        province:
          type: string
        station:
          type: string
        salesManager:
          type: string
        invoiceDate:
          type: string
        amountWithTax:
          type: number
        amountWithoutTax:
          type: number
        taxRate:
          type: number
        expectedPaymentDate:
          type: ["string", "null"]
        expectedPaymentAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - customer
        - province
        - station
        - salesManager
        - invoiceDate
        - amountWithTax
        - amountWithoutTax
        - taxRate
    UpdateInvoiceBody:
      type: object
      properties:
        contractId:
          type: ["integer", "null"]
        contractNo:
          type: ["string", "null"]
        invoiceNo:
          type: ["string", "null"]
        customer:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        station:
          type: ["string", "null"]
        salesManager:
          type: ["string", "null"]
        invoiceDate:
          type: ["string", "null"]
        amountWithTax:
          type: ["number", "null"]
        amountWithoutTax:
          type: ["number", "null"]
        taxRate:
          type: ["number", "null"]
        expectedPaymentDate:
          type: ["string", "null"]
        expectedPaymentAmount:
          type: ["number", "null"]
        actualPaymentDate:
          type: ["string", "null"]
        actualPaymentAmount:
          type: ["number", "null"]
        voidDate:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    DimensionStat:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        count:
          type: integer
      required:
        - name
        - amount
        - count
...