# lib/api-zod/src/generated/types/createContractBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Management System API for Wind/Solar Energy Forecasting Company
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContractBody {
  contractNo: string;
  contractName: string;
  contractType: string;
  status: string;
  customer: string;
  province: string;
  group: string;
  station: string;
  salesManager: string;
  /** @nullable */
  signDate?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  amountWithTax: number;
  amountWithoutTax: number;
  productType: string;
  /** @nullable */
  workOrderNo?: string | null;
  /** @nullable */
  afterSaleNo?: string | null;
  /** @nullable */
  notes?: string | null;
  isSpecial?: boolean;
}