# lib/api-zod/src/generated/types/createInvoiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Management System API for Wind/Solar Energy Forecasting Company
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInvoiceBody {
  /** @nullable */
  contractId?: number | null;
  /** @nullable */
  contractNo?: string | null;
  /** @nullable */
  invoiceNo?: string | null;
  customer: string;
  province: string;
  station: string;
  salesManager: string;
  invoiceDate: string;
  amountWithTax: number;
  amountWithoutTax: number;
  taxRate: number;
  /** @nullable */
  expectedPaymentDate?: string | null;
  /** @nullable */
  expectedPaymentAmount?: number | null;
  /** @nullable */
  notes?: string | null;
}